Moved chart data logic to client on TopPoolsChart
